Fix: RoutineManager 생성 및 함수와 매개변수명 변경
